Refs #21082. Move data-sharing related fields to `BaseReader`.
